fix exception when using this method on sequences with more then one period

+++ src/Sequence.php
@@ -226,7 +226,7 @@ final class Sequence implements ArrayAccess, Countable, IteratorAggregate, JsonS
      */
     public function getTotalTimestampInterval(): float
     {
-        $func = function (int $carry, Period $interval): float {
+        $func = function (float $carry, Period $interval): float {
             return $carry + $interval->getTimestampInterval();
         };
